Start refactoring MinCostFlow problems into dispatch / transmission (tests failing)
