+++ src/simulations/nonsequentialnetworkflow.jl
@@ -51,7 +51,7 @@ function assess!(
 
     statesampler = sampler(
         SystemInputStateDistribution(cache.system, t, copperplate=false))
-    flowproblem = FlowProblem(cache.simulationspec, cache.system)
+    flowproblem = TransmissionProblem(cache.system)
     outputsample = SystemOutputStateSample{L,T,P}(
         cache.system.interfaces.regions_from,
         cache.system.interfaces.regions_to,
@@ -60,37 +60,8 @@ function assess!(
     for i in 1:cache.simulationspec.nsamples
         rand!(cache.rngs[thread], flowproblem, statesampler)
         solveflows!(flowproblem)
-        update!(cache.simulationspec, outputsample, flowproblem)
+        update!(outputsample, flowproblem)
         update!(acc, outputsample, t, i)
-    end
-
-end
-
-function update!(
-    simulationspec::NonSequentialNetworkFlow,
-    sample::SystemOutputStateSample{L,T,P},
-    fp::FlowProblem
-) where {L,T<:Period,P<:PowerUnit}
-
-    nregions = length(sample.regions)
-    ninterfaces = length(sample.interfaces)
-
-    # Save gen available, gen dispatched, demand, demand served for each region
-    for i in 1:nregions
-        node = fp.nodes[i]
-        surplus_edge = fp.edges[2*ninterfaces + i]
-        shortfall_edge = fp.edges[2*ninterfaces + nregions + i]
-        sample.regions[i] = RegionResult{L,T,P}(
-            node.injection, surplus_edge.flow, shortfall_edge.flow)
-    end
-
-    # Save flow available, flow for each interface
-    for i in 1:ninterfaces
-        forwardedge = fp.edges[i]
-        forwardflow = forwardedge.flow
-        reverseflow = fp.edges[ninterfaces+i].flow
-        flow = forwardflow > reverseflow ? forwardflow : -reverseflow
-        sample.interfaces[i] = InterfaceResult{L,T,P}(forwardedge.limit, flow)
     end
 
 end
