+++ src/simulations/sequentialnetworkflow.jl
@@ -123,7 +123,7 @@ function assess!(
     # Initialize storage devices as empty
     fill!(stors_energy, 0)
 
-    flowproblem = FlowProblem(cache.simulationspec, cache.system)
+    flowproblem = TransmissionDispatchProblem(cache.system)
 
     # Main simulation loop
     for t in 1:N
@@ -156,141 +156,6 @@ function assess!(
         update!(cache.simulationspec, outputsample, flowproblem)
         update!(acc, outputsample, t, i)
 
-    end
-
-end
-
-function update_flownodes!(
-    flowproblem::FlowProblem,
-    t::Int,
-    loads::Matrix{Int},
-    genranges::Vector{Tuple{Int,Int}},
-    gens::Generators,
-    gens_available::Vector{Bool},
-    storranges::Vector{Tuple{Int,Int}},
-    stors::Storages,
-    stors_available::Vector{Bool},
-    stors_energy::Vector{Int},
-)
-
-    nregions = length(genranges)
-    slacknode = flowproblem.nodes[end]
-
-    for r in 1:nregions
-
-        region_node = flowproblem.nodes[r]
-        region_dischargenode = flowproblem.nodes[nregions + r]
-        region_chargenode = flowproblem.nodes[2*nregions + r]
-
-        # Update generators
-        gen_range = genranges[r]
-        region_gensurplus =
-            available_capacity(gens_available, gens, gen_range, t) - loads[r, t]
-        updateinjection!(region_node, slacknode, region_gensurplus)
-
-        # Update storages
-        stor_range = storranges[r]
-        charge_capacity, discharge_capacity = available_storage_capacity(
-            stors_available, stors_energy, stors, stor_range, t)
-        updateinjection!(region_chargenode, slacknode, -charge_capacity)
-        updateinjection!(region_dischargenode, slacknode, discharge_capacity)
-
-    end
-
-end
-
-function update_flowedges!(
-    flowproblem::FlowProblem,
-    t::Int,
-    lineranges::Vector{Tuple{Int,Int}},
-    lines::Lines,
-    lines_available::Vector{Bool}
-) where {V <: Real}
-
-    ninterfaces = length(lineranges)
-
-    for i in 1:ninterfaces
-
-        interface_forwardedge = flowproblem.edges[i]
-        interface_backwardedge = flowproblem.edges[ninterfaces + i]
-        line_range = lineranges[i]
-
-        interface_capacity_forward, interface_capacity_backward =
-            available_capacity(lines_available, lines, line_range, t)
-
-        updateflowlimit!(interface_forwardedge, interface_capacity_forward)
-        updateflowlimit!(interface_backwardedge, interface_capacity_backward)
-
-    end
-
-end
-
-function update_energy!(
-    stors_energy::Vector{Int},
-    t::Int,
-    storranges::Vector{Tuple{Int,Int}},
-    stors::Storages,
-    stors_available::Vector{Bool},
-    flowproblem::FlowProblem,
-    ninterfaces::Int
-)
-
-    nregions = length(storranges)
-    nstors = length(stors)
-
-    for r in 1:nregions
-
-        region_discharge = flowproblem.edges[2*ninterfaces + nregions + r].flow
-        region_charge = flowproblem.edges[2*ninterfaces + 3*nregions + r].flow
-
-        storrange = storranges[r]
-
-        if region_charge > 0
-
-            charge_storage!(
-                stors_available, stors_energy,
-                region_charge, stors, storrange, t)
-
-        elseif region_discharge > 0
-
-            discharge_storage!(
-                stors_available, stors_energy,
-                region_discharge, stors, storrange, t)
-
-        end
-
-    end
-
-end
-
-#TODO: This function is very similar to the NonSequentialNetworkFlow
-#      equivalent - refactor to share code?
-function update!(
-    simulationspec::SequentialNetworkFlow,
-    outputsample::SystemOutputStateSample{L,T,P},
-    flowproblem::FlowProblem
-) where {L,T<:Period,P<:PowerUnit}
-
-    nregions = length(outputsample.regions)
-    ninterfaces = length(outputsample.interfaces)
-
-    # Save gen available, gen dispatched, demand, demand served for each region
-    for i in 1:nregions
-        node = flowproblem.nodes[i]
-        surplus_edge = flowproblem.edges[2*ninterfaces + i]
-        shortfall_edge = flowproblem.edges[2*ninterfaces + 5*nregions + i]
-        outputsample.regions[i] = RegionResult{L,T,P}(
-            node.injection, surplus_edge.flow, shortfall_edge.flow)
-    end
-
-    # Save flow available, flow for each interface
-    for i in 1:ninterfaces
-        forwardedge = flowproblem.edges[i]
-        forwardflow = forwardedge.flow
-        reverseflow = flowproblem.edges[ninterfaces+i].flow
-        flow = forwardflow > reverseflow ? forwardflow : -reverseflow
-        outputsample.interfaces[i] =
-            InterfaceResult{L,T,P}(forwardedge.limit, flow)
     end
 
 end
